Adding a quick test of the metric manager clear functionality for coverage
