+++ tests/utils/metric_test.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import torch
 
-from fl4health.utils.metrics import F1, ROC_AUC, Accuracy, BalancedAccuracy, BinarySoftDiceCoefficient, MetricManager
+from fl4health.utils.metrics import F1, ROC_AUC, Accuracy, BalancedAccuracy, BinarySoftDiceCoefficient
 
 
 def test_accuracy_metric() -> None:
@@ -225,40 +225,3 @@ def test_metric_accumulation() -> None:
     # Accumulating the batches together results in recalls of (1.0, 1/5, 5/7) for 0, 1, 2 classes, these are then
     # averaged over the number of classes giving the correct balanced accuracy for the whole
     assert pytest.approx(acc_m_balanced_accuracy, abs=0.00001) == (1.0 + 1.0 / 5.0 + 5.0 / 7.0) / 3.0
-
-
-def test_metric_manager() -> None:
-    logits1 = torch.Tensor(
-        [
-            [0.8, 0.05, 0.15],
-            [0.88, 0.06, 0.06],
-            [0.1, 0.3, 0.6],
-            [0.4, 0.1, 0.5],
-            [0.1, 0.6, 0.3],
-        ]
-    )
-    target1 = torch.Tensor([0, 0, 2, 0, 2])
-
-    logits2 = torch.Tensor(
-        [
-            [0.4, 0.5, 0.1],
-            [0.1, 0.2, 0.7],
-            [0.3, 0.3, 0.4],
-            [0.75, 0.15, 0.1],
-            [0.1, 0.6, 0.3],
-        ]
-    )
-    target2 = torch.Tensor([1, 2, 2, 0, 1])
-
-    logits_list = [logits1, logits2]
-    target_list = [target1, target2]
-
-    mm = MetricManager([F1(), Accuracy()], "test")
-
-    for logits, target in zip(logits_list, target_list):
-        preds = {"prediction": logits}
-        mm.update(preds, target)
-    metrics = mm.compute()
-
-    assert metrics["test - prediction - F1 score"] == pytest.approx(0.80285714285, abs=0.00001)
-    assert metrics["test - prediction - accuracy"] == 0.8
